Remove redundant resize-window padding calculation
The "resize-window" signal handler incorrectly attempts to compute the
number of columns and rows. The 'width' and 'height' parameters are
in units of columns and rows while the 'pad_w' and 'pad_h' parameters
are pixels. Furthermore, a slightly different version of this
calculation is already performed in roxterm_set_vte_size.

Remove the incorrect calculation and simply pass the original resize
parameters on.

Signed-off-by: Perry Hung <iperry@gmail.com>